kernel: printk: Fix userspace log filter
The condition for the userspace log filter was returning "ret" instead of "len" which
killed the function before calling printk_emit(), resulting in producing no userspace
logs at all whatsoever. Fix the return check and flesh it out as a concrete condition
rather than a nested condition and replace the use of strncmp() with strstr() since
it's much simpler and doesn't require character numbers to track to compare the
strings while allowing us to compare any portion of a string with the full string.

Fixes: ("kernel: printk: Filter out some userspace logs")
Change-Id: a4103065b4d3b7603b79fb0d15148141d44a44b7
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: TogoFire <togofire@mailfence.com>